Add wifi ssid ble attribute (#29)
* Add wifi ssid ble attribute

Adds the wifi ssd attribute back in. This is a read only attribute
that notifies (when enabled) changes to the connected wifi ssid. If
wifi is not connected the ssid will be the empty string.

Also exposes a `gatway_config wifi online` cli command to show what
config thinks the currently connected wifi ssid is.